moved add data to database context
